requirements.txt
alembic==1.15.2
aniso8601==10.0.0
attrs==25.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.5
pytz==2025.2
referencing==0.36.2
rpds-py==0.24.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
Werkzeug==3.1.3
